feat(QueryEditor): execute query with selected text
