+++ src/containers/Tenant/ObjectGeneral/ObjectGeneral.tsx
@@ -18,7 +18,7 @@ import './ObjectGeneral.scss';
 const b = cn('object-general');
 
 interface ObjectGeneralProps {
-    type?: EPathType;
+    type: EPathType;
     additionalTenantProps?: AdditionalTenantsProps;
     additionalNodesProps?: AdditionalNodesProps;
 }
