+++ src/containers/Tenant/Query/Query.tsx
@@ -19,8 +19,8 @@ const b = block('ydb-query');
 
 interface QueryProps {
     theme: string;
-    path?: string;
-    type?: EPathType;
+    path: string;
+    type: EPathType;
 }
 
 export const Query = (props: QueryProps) => {
