+++ src/types/store/executeQuery.ts
@@ -3,16 +3,11 @@ import {
     changeUserInput,
     saveQueryToHistory,
     goToPreviousQuery,
-    setMonacoHotKey,
     goToNextQuery,
     setTenantPath,
-    MONACO_HOT_KEY_ACTIONS,
 } from '../../store/reducers/executeQuery';
 import type {ApiRequestAction} from '../../store/utils';
-import type {ValueOf} from '../common';
 import type {IQueryResult, QueryError, QueryErrorResponse} from './query';
-
-export type MonacoHotKeyAction = ValueOf<typeof MONACO_HOT_KEY_ACTIONS>;
 
 export interface QueryInHistory {
     queryText: string;
@@ -24,10 +19,9 @@ export interface ExecuteQueryState {
     input: string;
     history: {
         // String type for backward compatibility
-        queries: (QueryInHistory | string)[];
+        queries: QueryInHistory[];
         currentIndex: number;
     };
-    monacoHotKey: null | MonacoHotKeyAction;
     tenantPath?: string;
     data?: IQueryResult;
     stats?: IQueryResult['stats'];
@@ -42,7 +36,6 @@ export type ExecuteQueryAction =
     | ReturnType<typeof goToPreviousQuery>
     | ReturnType<typeof changeUserInput>
     | ReturnType<typeof saveQueryToHistory>
-    | ReturnType<typeof setMonacoHotKey>
     | ReturnType<typeof setTenantPath>;
 
 export interface ExecuteQueryStateSlice {
